New input format (Vertical) implemented based on #6
The new vertical imput format handling has been implemented for training
and tagging.

To choose the type of the desired input and output format, use the -if
and -of command line options. The available parameters, are: 'ord' for
the classic, ordinary format and 'vert' for the new vertical one.

The program is able to handle comments in the input files from now
between a <> tag. The tag cannot contains whitespaces and must separated
from the other content as the tokens are in the actual format. In case
of ordinary, with a whitespace, in the vertical the comment it must be a
line.  The comments will copied into the output file to the same
position where it was in the input if the input and output formats are
equal.