Mishandled character casings at lemma decoding #3
GeneralizedLemmaTransformation and SuffixLemmaTransformation wired out
to CLI with the threshold of GeneralizedLemmaTransformation. The new
options:

-l <class> where class can be "suffix" or "generalized". The default is
still the SuffixLemmaTransformation.

-lt <length> where length is the desired treshold for the
GeneralizedLemmaTransformation's class decode function. The default is
2.

When the selected Transformation class or threshold doesn't
match the defaults from the model file, a warning will appear on the
console and the defaults will used.

Merged the common methods and steps of Generalized and Suffix
Transformation classes into their Abstract ancestor (encode and
checkCasing functions and the common part of the decode function).

Created a new common representation for both classes: Pair<String,
Integer>, where:

String: the parts of the token which needed to assembly it from the
lemma.

Integer a ten digit long integer code, generated by shifting the
parameters with the static variables of the AbstractLemmaTransformation
class, where:
*    [1] casing, can be -1,0,1
*    [3] tag
*    [2] removeStart
*    [2] removeEnd
*    [2] addEnd

Implemented the constructor and the getter functions for this code.

Reworked the testing method of the LemmaTransformation classes.